fix: mask checkpoint storage secrets in gc logs
We already mask secrets in Trial logs, but we missed the secrets in
checkpoint gc logs.
